requirements.txt
numpy==1.11.2
scipy==0.18.1
pydicom==0.9.9
tabulate==0.7.5
